+++ src/components/Avatar/AvatarBase/index.ts
@@ -1,0 +1,1 @@
+export { AvatarBase, type AvatarBaseProps, type AvatarSize } from "./AvatarBase"
