+++ src/components/Avatar/AvatarGroup/README.md
